[update] fix intro

+++ intro/byteorder.c
@@ -18,7 +18,7 @@ main(int argc, char **argv)
 		else
 			printf("unknown\n");
 	} else
-		printf("sizeof(short) = %d\n", sizeof(short));
+		printf("sizeof(short) = %d\n",(int)sizeof(short));
 
 	exit(0);
 }
